SliderWidget: Allow float for autoplayInterval
Most (if not all) slider libraries offer a timer in milliseconds.

+++ src/Widgets/SliderWidget/SliderWidgetClass.ts
@@ -3,7 +3,7 @@ import { provideWidgetClass } from 'scrivito'
 export const SliderWidget = provideWidgetClass('SliderWidget', {
   attributes: {
     autoplay: 'boolean',
-    autoplayInterval: 'integer',
+    autoplayInterval: 'float',
     controls: 'boolean',
     margin: [
       'enum',
